Move test helper functions to TestUtils
